save, update +

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -50,7 +50,7 @@ public class AdminController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id")int id) {
+    public String delete(@PathVariable("id")Long id) {
         userService.delete(id);
         return "redirect:/admin";
     }
